requirements.txt
binance-connector==3.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
kucoin-python==1.0.11
pybit==5.3.0
pycryptodome==3.18.0
requests==2.31.0
urllib3==2.0.3
websocket-client==1.6.0
websockets==11.0.3
telebot~=0.0.5
python-dotenv~=1.0.0